# Simplilearn2/Phase2Section5.8/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Phase2Section5.7",
    "version": "1.0"
  },
  "paths": {
    "/api/Students": {
      "get": {
        "tags": [
          "Students"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Students"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Students/{id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "StudentModel": {
        "required": [
          "address",
          "contactEmail",
          "course",
          "grades",
          "name"
        ],
        "type": "object",
        "properties": {
          "studentID": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "address": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "contactEmail": {
            "type": "string",
            "format": "email"
          },
          "course": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string"
          },
          "grades": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}